+++ CHANGELOG.md
@@ -22,6 +22,8 @@
 * The following attributes have been renamed in the `ViewMenu` class
   * `disable_buttons_on_timeout` → `disable_items_on_timeout`
   * `remove_buttons_on_timeout` → `remove_items_on_timeout`
+#### Bug Fixes
+* Fixed an issue where `ViewButton` link buttons would be removed even after setting is as persistent
 #### New Features
 ##### ReactionMenu & ViewMenu
 * Pagination is no longer limited to just embeds or text. The normal embed menu can now paginate with embeds, text, as well as files. With this, method `.add_page()` has two new parameters